meta: fix all remaining warnings

+++ options/glibc/generic/shadow-stubs.cpp
@@ -13,7 +13,7 @@ int putspent(const struct spwd *sp, FILE *f) {
 	return fprintf(f, "%s:%s:%.*ld:%.*ld:%.*ld:%.*ld:%.*ld:%.*ld:%.*lu\n",
 		str(sp->sp_namp), str(sp->sp_pwdp), NUM(sp->sp_lstchg),
 		NUM(sp->sp_min), NUM(sp->sp_max), NUM(sp->sp_warn),
-		NUM(sp->sp_inact), NUM(sp->sp_expire), NUM(sp->sp_flag)) < 0 ? -1 : 0;
+		NUM(sp->sp_inact), NUM(sp->sp_expire), NUM((int)sp->sp_flag)) < 0 ? -1 : 0;
 }
 #undef NUM
